create and update operation

+++ payloads/operation.go
@@ -1,11 +1,11 @@
 package payloads
 
 type Operation []struct {
-	AccountId   uint    `json:"TR_ID"`
-	OrderNumber float32 `json:"POS_NUM"`
+	AccountId   int     `json:"TR_ID"`
+	OrderNumber int     `json:"POS_NUM"`
 	Symbol      string  `json:"SYMBOL"`
-	OrderType   uint    `json:"ORD_TYPE"`
-	Price       float32 `json:"PRICE"`
-	StopLoss    float32 `json:"SL"`
-	TakeProfit  float32 `json:"TP"`
+	OrderType   int     `json:"ORD_TYPE"`
+	Price       float64 `json:"PRICE"`
+	StopLoss    float64 `json:"SL"`
+	TakeProfit  float64 `json:"TP"`
 }
